Refactor client interface to client package. Improve javadoc
