+++ src/test/java/org/dbos/apiary/PostgresTests.java
@@ -8,8 +8,8 @@ import org.dbos.apiary.procedures.postgres.retwis.*;
 import org.dbos.apiary.utilities.ApiaryConfig;
 import org.dbos.apiary.worker.ApiaryNaiveScheduler;
 import org.dbos.apiary.worker.ApiaryWorker;
-import org.dbos.apiary.worker.ApiaryWorkerClient;
-import org.dbos.apiary.worker.InternalApiaryWorkerClient;
+import org.dbos.apiary.client.ApiaryWorkerClient;
+import org.dbos.apiary.client.InternalApiaryWorkerClient;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
