+++ src/test/java/org/dbos/apiary/VoltDBTests.java
@@ -8,8 +8,8 @@ import org.dbos.apiary.utilities.ApiaryConfig;
 import org.dbos.apiary.voltdb.VoltConnection;
 import org.dbos.apiary.worker.ApiaryNaiveScheduler;
 import org.dbos.apiary.worker.ApiaryWorker;
-import org.dbos.apiary.worker.ApiaryWorkerClient;
-import org.dbos.apiary.worker.InternalApiaryWorkerClient;
+import org.dbos.apiary.client.ApiaryWorkerClient;
+import org.dbos.apiary.client.InternalApiaryWorkerClient;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
